requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.6
django-autoslug==1.9.8
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-crispy-forms==1.11.0
django-filter==2.4.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
kombu==5.0.2
Pillow==8.1.0
prompt-toolkit==3.0.16
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
